optimized #1449 codebase
