+++ lib/browser/index.js
@@ -43,11 +43,20 @@ riot.tag = function(name, html, css, attrs, fn) {
   }
   if (css) {
     if (isFunction(css)) fn = css
-    else if (tagStyles) tagStyles.add(css)
+    else if (styleManager) styleManager.add(css)
+
+    // make the style node available in the riot API
+    riot.styleNode = styleManager.styleNode
   }
   __tagImpl[name] = { name: name, tmpl: html, attrs: attrs, fn: fn }
   return name
 }
+
+/**
+ * This attribute will be created only if there will be css injected into the DOM
+ * @type { DOM Object }
+ */
+riot.styleNode = null
 
 /**
  * Create a new riot tag implementation (for use by the compiler)
@@ -60,7 +69,7 @@ riot.tag = function(name, html, css, attrs, fn) {
  * @returns { String } name/id of the tag just created
  */
 riot.tag2 = function(name, html, css, attrs, fn, bpair) {
-  if (css && tagStyles) tagStyles.add(css)
+  if (css && styleManager) styleManager.add(css)
   //if (bpair) riot.settings.brackets = bpair
   __tagImpl[name] = { name: name, tmpl: html, attrs: attrs, fn: fn }
   return name
@@ -112,7 +121,7 @@ riot.mount = function(selector, tagName, opts) {
   // ----- mount code -----
 
   // inject styles into DOM
-  if (tagStyles) tagStyles.inject()
+  if (styleManager) styleManager.inject()
 
   if (typeof tagName === T_OBJECT) {
     opts = tagName
