+++ lib/browser/tag/util.js
@@ -452,68 +452,6 @@ function startsWith(src, str) {
 }
 
 /**
- * Function needed to inject in runtime the custom tags css
- */
-var tagStyles = (function() {
-
-  if (!window) return // skip injection on the server
-
-  // create the style node
-  var styleNode = mkEl('style'),
-    placeholder = $('style[type=riot]')
-
-  setAttr(styleNode, 'type', 'text/css')
-
-  // inject the new node into the DOM -- in head
-  if (placeholder) {
-    placeholder.parentNode.replaceChild(styleNode, placeholder)
-    placeholder = null
-  }
-  else document.getElementsByTagName('head')[0].appendChild(styleNode)
-
-  // extends the <style> node for dynamic CSS
-  styleNode.stylesToInject = ''
-  styleNode.needUpdate = false
-
-  // make the style node available in the riot API
-  riot.styleNode = styleNode
-
-  styleNode.updateStyles = function() {
-    if (styleNode.stylesToInject) {
-      var css = styleNode.stylesToInject
-      if (styleNode.parser) {
-        css = styleNode.parser(css)
-      }
-      if (styleNode.styleSheet) styleNode.styleSheet.cssText = css
-      else styleNode.innerHTML = css
-    }
-  }
-
-  return {
-    /**
-      * Save a tag style to be later injected into DOM
-      * @param   { String } css [description]
-      */
-    add: function(css) {
-      styleNode.stylesToInject += css
-      styleNode.needUpdate = true
-    },
-
-    /**
-      * Inject all previously saved tag styles into DOM
-      * innerHTML seems slow: http://jsperf.com/riot-insert-style
-      */
-    inject: function() {
-      if (styleNode.needUpdate) {
-        styleNode.updateStyles()
-        styleNode.needUpdate = false
-      }
-    }
-  }
-
-})()
-
-/**
  * requestAnimationFrame function
  * Adapted from https://gist.github.com/paulirish/1579671, license MIT
  */
