+++ lib/riot.js
@@ -8,6 +8,7 @@ import '../node_modules/riot-route/dist/riot.route.js'
 import '../node_modules/riot-tmpl/dist/riot.tmpl.js'
 import 'browser/tag/mkdom'
 import 'browser/tag/each'
+import 'browser/tag/styleManager'
 import 'browser/tag/parse'
 import 'browser/tag/tag'
 import 'browser/tag/update'
